requirements.txt
azure-ai-formrecognizer>=3.2.0
PyMuPDF>=1.22.5
Pillow
img2pdf